apps/analyzer/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chessbot-analyzer"
version = "0.1.0"
description = "Chess game analysis pipeline with Stockfish engine"
authors = [{ name = "Chess Autopost Team" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=1.10,<2",
    "python-chess>=1.999",
    "typer[all]>=0.9.0",
    "loguru>=0.7.0",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=1.4.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "zstandard>=0.22.0",
    "tqdm>=4.66.0",
    "orjson>=3.9.0"  # optional, remove if you prefer stdlib json
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "mypy>=1.0.0",
    "types-requests>=2.31.0.10"
]

[project.scripts]
chessbot = "chessbot_analyzer.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["chessbot_analyzer*"]

[tool.setuptools]
py-modules = []

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
